server/requirements.txt
fastapi==0.95.0
uvicorn==0.23.0
asyncpg==0.27.0
redis==4.4.0
python-dotenv==0.21.0
pydantic==1.10.4
datetime==4.3
typing-extensions==4.5.0
